refactor : 랜덤 숫자 생성 로직을 인터페이스로 추상화
