+++ src/main/java/domain/NumberGenerator.java
@@ -1,0 +1,6 @@
+package domain;
+
+@FunctionalInterface
+public interface NumberGenerator {
+    int generate();
+}
